refactor(ai): remove inspector and legacy chat paths
