+++ src/features/workspaces/components/ai-chat/constants.ts
@@ -1,10 +1,2 @@
-export {
-	DEFAULT_WORKSPACE_AI_CHAT_MODEL_ID,
-	WORKSPACE_AI_CHAT_MODELS as AI_CHAT_MODELS,
-} from "#/features/workspaces/ai/models";
-export {
-	AI_CHAT_ATTACHMENT_ACCEPT,
-	AI_CHAT_MAX_FILE_SIZE,
-	AI_CHAT_MAX_FILES,
-	WORKSPACE_AI_CHAT_ATTACHMENT_POLICY,
-} from "#/features/workspaces/ai/chat-attachment-policy";
+export { DEFAULT_WORKSPACE_AI_CHAT_MODEL_ID } from "#/features/workspaces/ai/models";
+export { WORKSPACE_AI_CHAT_ATTACHMENT_POLICY } from "#/features/workspaces/ai/chat-attachment-policy";
